requirements.txt
python-dotenv==1.0.0
gradio==3.41.0